init: qcom.rc: Adjust the console log threshold for the kernel
This change sets the console loglevel to < KERN_INFO and
set the default message loglevel to KERN_INFO.

One implication is 'printk()' with no specific priority
will not show up on the console by default.

This is done to prevent large volumes of info prints to the
serial console, which on our systems causes watchdog bites.

Change-Id: I0fec065443c5e362ec2dc0c0cf6bab27d94150d8